Improves camera auto adjustment
Adds pitch gravity to make the camera respond to ground normal.

Refactors yaw gravity to use dot product for a smoother effect.

Introduces a general gravity delay and uses elapsed time
since input was released to gate gravity effects.

Moves the camera to inactive in the Player prefab and enables the
terrain and disables the static flag.